not understanding why non-hpc-pod does not execute when on top#2
